requirements.txt
python-telegram-bot==20.7
psutil==5.9.6
python-dotenv==1.0.0
netifaces==0.11.0
